# api/openapi.json
{
  "components": {
    "responses": {
...
      "NotAcceptable": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`not_acceptable`: `Accept` excludes application/vnd.msgpack."
      },
...
      "Restricted": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`account_deletion_pending`: the account is scheduled for deletion and this session is export-only; `password_change_required`: the password was reset and must be changed with `PATCH /me` first."
      },
      "Unauthorized": {
        "content": {
          "application/problem+msgpack": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "description": "`unauthorized`: missing, invalid, expired or revoked access token (signed out, device removed, account disabled)."
      },
...
    },
    "schemas": {
...
      "DocumentList": {
        "description": "Documents.",
        "properties": {
          "items": {
            "description": "By name.",
            "items": {
              "$ref": "#/components/schemas/DocumentSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "DocumentStatus": {
        "description": "Where a document is, per its newest custody event.",
        "enum": [
          "stored",
          "checked-out",
          "with-third-party",
          "lost",
          "destroyed"
        ],
        "type": "string"
      },
...
      "EntityKind": {
        "description": "Kind of an entity.",
        "enum": [
          "person",
          "company",
          "document",
          "place"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "description": "Short-lived access token (PLAN §8).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Strata HTTP API. Every body is MessagePack (application/vnd.msgpack); errors are RFC 7807 problem details (application/problem+msgpack). See docs/WIRE_FORMAT.md.",
    "title": "Strata API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "operationId": "list_documents",
        "parameters": [
          {
            "description": "Name or alias, any script.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only documents in this place or any place nested inside it.",
            "in": "query",
            "name": "place",
            "required": false,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          },
          {
            "description": "Only documents this person holds now.",
            "in": "query",
            "name": "holder",
            "required": false,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          },
          {
            "description": "Only this status.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DocumentStatus"
            }
          },
          {
            "description": "Only documents expiring strictly before this date.",
            "in": "query",
            "name": "expiring_before",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentList"
                }
              }
            },
            "description": "Documents."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Documents, filtered (`place` includes nested places).",
        "tags": [
          "entities"
        ]
      },
      "post": {
        "operationId": "create_document",
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Created."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`duplicate_candidates`."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Create a document (duplicate check unless `force`).",
        "tags": [
          "entities"
        ]
      }
    },
...
    "/api/v1/entities": {
      "get": {
        "operationId": "list_entities",
        "parameters": [
          {
            "description": "Only this kind.",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EntityKind"
            }
          },
          {
            "description": "Name or alias, any script (normalised, substring or trigram match).",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only entities with this tag.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum results (1–500, default 100).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/EntityList"
                }
              }
            },
            "description": "Entities."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "422": {
            "$ref": "#/components/responses/InvalidParameter"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Entities of any or one kind, matching `q` in any script.",
        "tags": [
          "entities"
        ]
      },
      "post": {
        "operationId": "create_entity",
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            },
            "description": "Created; one `user: create <path>` commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`duplicate_candidates` (resend with `force`)."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Create a person, company, document or place (duplicate check unless `force`).",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/v1/entities/{id}": {
      "get": {
        "operationId": "get_entity",
        "parameters": [
          {
            "description": "Entity ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            },
            "description": "The entity."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "The aggregated entity page.",
        "tags": [
          "entities"
        ]
      },
      "patch": {
        "operationId": "patch_entity",
        "parameters": [
          {
            "description": "Entity ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          },
          {
            "description": "Optional: the note's current version.",
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/vnd.msgpack": {
              "schema": {
                "$ref": "#/components/schemas/PatchEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Entity"
                }
              }
            },
            "description": "Updated in one commit."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "409": {
            "content": {
              "application/problem+msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "`duplicate_candidates` (new aliases), `version_conflict` or `path_taken`."
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidBody"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Edit user fields, aliases, tags or the name (contact fields are only ever user-written).",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/v1/entities/{id}/documents": {
      "get": {
        "operationId": "get_entity_documents",
        "parameters": [
          {
            "description": "Entity ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDocuments"
                }
              }
            },
            "description": "Related documents."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Documents the entity holds now, last handled, or that concern it.",
        "tags": [
          "entities"
        ]
      }
    },
...
    "/api/v1/entities/{id}/notes": {
      "get": {
        "operationId": "get_entity_notes",
        "parameters": [
          {
            "description": "Entity ID (ULID).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "ulid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.msgpack": {
                "schema": {
                  "$ref": "#/components/schemas/MentioningNotes"
                }
              }
            },
            "description": "Mentioning notes."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Restricted"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "summary": "Notes mentioning the entity, newest first, with the mention line.",
        "tags": [
          "entities"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}